+++ data_bridges_knots/client.py
@@ -28,10 +28,25 @@ from data_bridges_knots.endpoints.householdApi import (
     get_household_xlsform_definition,
 )
 from data_bridges_knots.endpoints.marketPricesApi import get_prices
-from data_bridges_knots.endpoints.marketsApi import (get_market_geojson_list, get_markets_list, get_markets_as_csv, get_nearby_markets )
+from data_bridges_knots.endpoints.marketsApi import (
+    get_market_geojson_list,
+    get_markets_as_csv,
+    get_markets_list,
+    get_nearby_markets,
+)
 
+from data_bridges_knots.endpoints.economicDataApi import get_economic_indicator_list
 
-from data_bridges_knots.helpers import get_adm0_code
+from data_bridges_knots.endpoints.globalOutlookApi import get_global_outlook
+
+from data_bridges_knots.endpoints.rpmeApi import (
+    get_rpme_base_data,
+    get_rpme_full_data,
+    get_rpme_output_values,
+    get_rpme_surveys,
+    get_rpme_variables,
+    get_rpme_xls_forms,
+)
 
 logname = "data_bridges_api_calls.log"
 logging.basicConfig(
@@ -220,314 +235,26 @@ class DataBridgesKnots:
         logger.debug("Token used: %s", token.__repr__())
         return configuration
 
-    def get_economic_indicator_list(
-        self,
-        page: Optional[int] = 1,
-        indicator_name: Optional[str] = "",
-        country_iso3: Optional[str] = "",
-        format: Optional[str] = "json",
-    ):
-        """
-        Returns the lists of indicators for which Vulnerability Analysis and Mapping - Economic and Market Analysis Unit has redistribution licensing from Trading Economics.
-
-        Args:
-            page (int, optional): Page number for paged results. Defaults to 1.
-            indicator_name (str, optional): Unique indicator name. Defaults to ''.
-            iso3 (str, optional): The code to identify the country. Must be a ISO-3166 Alpha 3 code. Defaults to ''.
-            format (str, optional): Output format: 'json' or 'csv'. Defaults to 'json'.
-
-        Returns:
-            pandas.DataFrame: A DataFrame containing the retrieved economic indicator data.
-        """
-        with data_bridges_client.ApiClient(
-            self._setup_configuration_and_authentication(self.config)
-        ) as api_client:
-            # Create an instance of the API class
-            api_instance = data_bridges_client.EconomicDataApi(api_client)
-
-            try:
-                # Returns the lists of indicators.
-                api_response = api_instance.economic_data_indicator_list_get(
-                    page=page,
-                    indicator_name=indicator_name,
-                    iso3=country_iso3,
-                    format=format,
-                    env=self.env,
-                )
-                logger.info(
-                    "The response of EconomicDataApi->economic_data_indicator_list_get:\n"
-                )
-                df = pd.DataFrame([item.to_dict() for item in api_response.items])
-                df = df.replace({np.nan: None})
-                return df
-                return api_response
-            except Exception as e:
-                logger.error(
-                    "Exception when calling EconomicDataApi->economic_data_indicator_list_get: %s",
-                    e,
-                )
-                raise
-
-
-    def get_global_outlook(
-        self,
-        data_type: Literal["country_latest", "global_latest", "regional_latest"],
-        page: Optional[int] = None,
-    ) -> pd.DataFrame:
-        """Retrieves data from the Global Outlook API.
-
-        The Global Outlook API provides access to WFP’s forward-looking analysis and
-        aggregated insights at different geographical levels, including country,
-        regional, and global summaries.
-
-        Args:
-            data_type (str): The type of Global Outlook data to retrieve. Must be one of:
-                - 'country_latest': Latest data at country level
-                - 'global_latest': Latest global aggregated data
-                - 'regional_latest': Latest data aggregated by region
-            page (int, optional): Page number for paginated results. Currently not used
-                for latest endpoints. Defaults to None.
-
-        Returns:
-            pd.DataFrame: DataFrame containing Global Outlook data for the selected scope.
-
-        Examples:
-            >>> client = DataBridgesShapes("data_bridges_api_config.yaml")
-            >>> # Get latest country-level outlook
-            >>> country_data = client.get_global_outlook("country_latest")
-            >>> # Get global outlook summary
-            >>> global_data = client.get_global_outlook("global_latest")
-            >>> # Get regional outlook data
-            >>> regional_data = client.get_global_outlook("regional_latest")
-
-        Raises:
-            ValueError: If data_type is not one of the allowed values
-            ApiException: If there is an error accessing the Global Outlook API
-        """
-
-        # Enter a context with an instance of the API client
-        with data_bridges_client.ApiClient(
-            self._setup_configuration_and_authentication(self.config)
-        ) as api_client:
-            # Create an instance of the API class
-            api_instance = data_bridges_client.GlobalOutlookApi(api_client)
-            env = (
-                self.env
-            )  # str | Environment.   * `prod` - api.vam.wfp.org   * `dev` - dev.api.vam.wfp.org (optional)
-
-            try:
-                if data_type == "country_latest":
-                    api_response = api_instance.global_outlook_country_latest_get(
-                        env=env
-                    )
-                elif data_type == "global_latest":
-                    api_response = api_instance.global_outlook_global_latest_get(
-                        env=env
-                    )
-
-                elif data_type == "regional_latest":
-                    api_response = api_instance.global_outlook_regional_latest_get(
-                        env=env
-                    )
-                else:
-                    raise ValueError(f"Invalid data_type: {data_type}")
-                logger.info(
-                    f"Successfully retrieved Global Outlook data for type: {data_type}"
-                )
-                return pd.DataFrame([item.to_dict() for item in api_response.items])
-
-            except Exception as e:
-                logger.error(
-                    "Exception when calling GlobalOutlookApi->%s: %s", data_type, e
-                )
-                raise
-
-
-
-        # TODO: Get the scope and test these functions
-        def get_rpme_base_data(
-            self, survey_id=None, page: Optional[int] = 1, page_size=20
-        ):
-            with data_bridges_client.ApiClient(
-                self._setup_configuration_and_authentication(self.config)
-            ) as api_client:
-                api_instance = data_bridges_client.RpmeApi(api_client)
-                env = self.env
-
-                try:
-                    api_response = api_instance.rpme_base_data_get(
-                        survey_id=survey_id, page=page, page_size=page_size, env=env
-                    )
-                    logger.info("Successfully retrieved RPME base data")
-                    df = pd.DataFrame([item.to_dict() for item in api_response.items])
-                    df = df.replace({np.nan: None})
-                    return df
-                except ApiException as e:
-                    logger.error(
-                        f"Exception when calling RpmeApi->rpme_base_data_get: {e}"
-                    )
-                    raise
-
-        # TODO: Get the scope and test these functions
-        def get_rpme_full_data(
-            self,
-            survey_id=None,
-            format: Optional[str] = "json",
-            page: Optional[int] = 1,
-            page_size=20,
-        ):
-            with data_bridges_client.ApiClient(
-                self._setup_configuration_and_authentication(self.config)
-            ) as api_client:
-                api_instance = data_bridges_client.RpmeApi(api_client)
-                env = self.env
-
-                try:
-                    api_response = api_instance.rpme_full_data_get(
-                        survey_id=survey_id,
-                        format=format,
-                        page=page,
-                        page_size=page_size,
-                        env=env,
-                    )
-                    logger.info("Successfully retrieved RPME full data")
-                    df = pd.DataFrame([item.to_dict() for item in api_response.items])
-                    df = df.replace({np.nan: None})
-                    return df
-                except ApiException as e:
-                    logger.error(
-                        f"Exception when calling RpmeApi->rpme_full_data_get: {e}"
-                    )
-                    raise
-
-        # TODO: Get the scope and test these functions
-        def get_rpme_output_values(
-            self,
-            page: Optional[int] = 1,
-            adm0_code=None,
-            survey_id=None,
-            shop_id=None,
-            market_id=None,
-            adm0_code_dots="",
-        ):
-            with data_bridges_client.ApiClient(
-                self._setup_configuration_and_authentication(self.config)
-            ) as api_client:
-                api_instance = data_bridges_client.RpmeApi(api_client)
-                env = self.env
-
-                try:
-                    api_response = api_instance.rpme_output_values_get(
-                        page=page,
-                        adm0_code=adm0_code,
-                        survey_id=survey_id,
-                        shop_id=shop_id,
-                        market_id=market_id,
-                        adm0_code_dots=adm0_code_dots,
-                        env=env,
-                    )
-                    logger.info("Successfully retrieved RPME output values")
-                    df = pd.DataFrame([item.to_dict() for item in api_response.items])
-                    df = df.replace({np.nan: None})
-                    return df
-                except ApiException as e:
-                    logger.error(
-                        f"Exception when calling RpmeApi->rpme_output_values_get: {e}"
-                    )
-                    raise
-
-        # TODO: Get the scope and test these functions
-        def get_rpme_surveys(
-            self, adm0_code=0, page: Optional[int] = 1, start_date=None, end_date=None
-        ):
-            with data_bridges_client.ApiClient(
-                self._setup_configuration_and_authentication(self.config)
-            ) as api_client:
-                api_instance = data_bridges_client.RpmeApi(api_client)
-                env = self.env
-
-                try:
-                    api_response = api_instance.rpme_surveys_get(
-                        adm0_code=adm0_code,
-                        page=page,
-                        start_date=start_date,
-                        end_date=end_date,
-                        env=env,
-                    )
-                    logger.info("Successfully retrieved RPME surveys")
-                    df = pd.DataFrame([item.to_dict() for item in api_response.items])
-                    df = df.replace({np.nan: None})
-                    return df
-                except ApiException as e:
-                    logger.error(
-                        f"Exception when calling RpmeApi->rpme_surveys_get: {e}"
-                    )
-                    raise
-
-        # TODO: Get the scope and test these functions
-        def get_rpme_variables(self, page: Optional[int] = 1):
-            with data_bridges_client.ApiClient(
-                self._setup_configuration_and_authentication(self.config)
-            ) as api_client:
-                api_instance = data_bridges_client.RpmeApi(api_client)
-                env = self.env
-
-                try:
-                    api_response = api_instance.rpme_variables_get(page=page, env=env)
-                    logger.info("Successfully retrieved RPME variables")
-                    df = pd.DataFrame([item.to_dict() for item in api_response.items])
-                    df = df.replace({np.nan: None})
-                    return df
-                except ApiException as e:
-                    logger.error(
-                        f"Exception when calling RpmeApi->rpme_variables_get: {e}"
-                    )
-                    raise
-
-        # TODO: Get the scope and test these functions
-        def get_rpme_xls_forms(
-            self, adm0_code=0, page: Optional[int] = 1, start_date=None, end_date=None
-        ):
-            with data_bridges_client.ApiClient(
-                self._setup_configuration_and_authentication(self.config)
-            ) as api_client:
-                api_instance = data_bridges_client.RpmeApi(api_client)
-                env = self.env
-
-                try:
-                    api_response = api_instance.rpme_xls_forms_get(
-                        adm0_code=adm0_code,
-                        page=page,
-                        start_date=start_date,
-                        end_date=end_date,
-                        env=env,
-                    )
-                    logger.info("Successfully retrieved RPME XLS forms")
-                    df = pd.DataFrame([item.to_dict() for item in api_response.items])
-                    df = df.replace({np.nan: None})
-                    return df
-                except ApiException as e:
-                    logger.error(
-                        f"Exception when calling RpmeApi->rpme_xls_forms_get: {e}"
-                    )
-                    raise
-
-        # Add this function to the DataBridgesShapes class
-
+    # TODO: IpcchApi
     def get_ipc_and_equivalent_data(self):
         pass
 
+    # TODO: HungerHotspotApi
     def get_hotpost_data(self):
         pass
 
     def get_aims_data(self):
         pass
-
-    def get_rpme_data(self):
-        pass
-
-    def get_cari_data(self):
-        pass
+    
+    # TODO: IncubationApi 
+    # FIXME: Get scopes to test this function
+    def get_cari_data(self, admin_level = "admin0"):
+        if admin_level == "admin0":
+            pass
+        elif admin_level == "admin1":
+            pass
+        else:
+            raise ValueError("admin_level must be either 'admin0' or 'admin1'")
 
 
 class DataBridgesShapes(DataBridgesKnots):
@@ -583,6 +310,19 @@ DataBridgesKnots.marketsApi.get_markets_list = get_markets_list
 DataBridgesKnots.marketsApi.get_markets_as_csv = get_markets_as_csv
 DataBridgesKnots.marketsApi.get_nearby_markets = get_nearby_markets
 
+# RPME Endpoints (RpmeApi)
+DataBridgesKnots.rpmeApi.get_rpme_base_data = get_rpme_base_data
+DataBridgesKnots.rpmeApi.get_rpme_full_data = get_rpme_full_data
+DataBridgesKnots.rpmeApi.get_rpme_output_values = get_rpme_output_values
+DataBridgesKnots.rpmeApi.get_rpme_surveys = get_rpme_surveys
+DataBridgesKnots.rpmeApi.get_rpme_variables = get_rpme_variables
+DataBridgesKnots.rpmeApi.get_rpme_xls_forms = get_rpme_xls_forms
+
+# Global Outlook
+DataBridgesKnots.globalOutlookApi.get_global_outlook = get_global_outlook
+
+# Economic Data
+DataBridgesKnots.economicDataApi.get_economic_indicator_list = get_economic_indicator_list  
 
 if __name__ == "__main__":
     pass
